==> dataset_preparation/0_preprocessing.ipynb <==
import os

print(os.path.abspath(os.curdir))

import pandas as pd

vto_df = pd.read_csv('train/parsed-cleansed.csv')
vto_df.to_excel('train/vto_data.xlsx', index=False)

### Merge the dataset
Merging the train and test set into one single file for annotations

import pandas as pd
import os

def load_and_process_data(set='train'):
    # Load data from a CSV file
    merged_data = pd.DataFrame()

    print("Loading training data...")
    files = os.listdir(set)
    for file in files:
        full_path = os.path.join(set, file)
        data = pd.read_excel(full_path)
        file_df = pd.DataFrame()
        file_df['set'] = [set] * len(data)
        file_df['source_file'] = [file] * len(data)
        file_df['message'] = data['message']
        merged_data = pd.concat([merged_data, file_df], ignore_index=True)
    
    # Display the first few rows of the dataframe
    print("First few rows of the dataset:")
    print(merged_data.head())
    
    # Display summary statistics of the dataframe
    print("\nSummary statistics of the dataset:")
    print(merged_data.describe())
    
    # Check for missing values in the dataframe
    print("\nMissing values in each column:")
    print(merged_data.isnull().sum())
    
    return merged_data

train_df = load_and_process_data('train')
train_df.head(5)

test_df = load_and_process_data('test')

merged_df = pd.concat([train_df, test_df], ignore_index=True)
merged_df.to_excel('merged_data.xlsx', index=False)
print("Merged DataFrame shape:", merged_df.shape)

print(merged_df.describe())

# Suppose the column containing the log messages is called 'message'
train_messages = set(train_df['message'].str.lower())
test_messages = set(test_df['message'].str.lower())

# Find overlaps
overlaps = train_messages.intersection(test_messages)
print("Overallping messages:\n", overlaps)
print(f"Number of overlapping messages: {len(overlaps)}")
print(f"Percentage overlap (relative to test set): {len(overlaps) / len(test_messages) * 100:.2f}%")

# Optionally, inspect the overlapping samples
overlap_df = test_df[test_df['message'].isin(overlaps)]
print(overlap_df.head())
